Controller 리팩토링 및 버그 수정
1. Service 레이어 변경에 따른 Controller 변경
2. 스크린, 배터리 사이즈 정수밖에 못받던 버그 수정
3. 카테고리 선택 안했을때 에러나던 버그 수정
4. 상품 추가시 년도와 함께 월정보도 추가되던 버그 수정

+++ stupping-mall/src/main/java/com/nerds/stuppingmall/domain/Notebook.java
@@ -28,11 +28,11 @@ public class Notebook {
 	private String cpuName;
 	private String gpuName;
 	private double weight;
-	private int screenSize;
+	private double screenSize;
 	private int ramSize;
 	private int ssdSize;
 	private int hddSize;
-	private int batterySize;
+	private double batterySize;
 	private String usage;
 	
 	@Override
