+++ stupping-mall/src/main/java/com/nerds/stuppingmall/dto/NotebookAddRequestDto.java
@@ -13,9 +13,9 @@ public class NotebookAddRequestDto {
 	private String cpuName;
 	private String gpuName;
 	private double weight;
-	private int screenSize;
+	private double screenSize;
 	private int ramSize;
 	private int ssdSize;
 	private int hddSize;
-	private int batterySize;
+	private double batterySize;
 }
